fixing x64 issues
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@984 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/util/impl/thread_win32.inl
@@ -588,7 +588,7 @@ public:
 		}
 	}
 
-	void bind(unsigned processor)
+	void bind(size_t processor)
 	{
 		affinity_ = bindThread(handle_, processor);
 	}
@@ -614,7 +614,7 @@ public:
 		Sleep(0);
 	}
 
-	static void bindCurrent(unsigned processor)
+	static void bindCurrent(size_t processor)
 	{
 		bindThread(GetCurrentThread(), processor);
 	}
